# Conflicts:
#	public/build/manifest.json

+++ database/migrations/2023_06_15_210220_create_user_logs_table.php
@@ -10,6 +10,9 @@ return new class () extends Migration {
      */
     public function up(): void
     {
+        if (!config('logging.api')) {
+            return;
+        }
         Schema::connection('logs')->create('user_logs', function (Blueprint $table) {
             $table->id();
             $table->integer('user_id')->unsigned();
@@ -27,6 +30,9 @@ return new class () extends Migration {
      */
     public function down(): void
     {
+        if (!config('logging.api')) {
+            return;
+        }
         Schema::dropIfExists('user_logs');
     }
 };
